+++ src/Curl.php
@@ -11,6 +11,7 @@ use Innmind\Http\{
     Request,
     Factory\Header\Factory,
 };
+use Innmind\Url\Url;
 use Innmind\TimeContinuum\{
     Clock,
     Period,
@@ -33,6 +34,7 @@ final class Curl implements Transport
         private Period $timeout,
         private \Closure $heartbeat,
         private bool $disableSSLVerification,
+        private ?Url $proxy,
     ) {
     }
 
@@ -44,6 +46,7 @@ final class Curl implements Transport
             $this->io,
             $request,
             $this->disableSSLVerification,
+            $this->proxy,
         );
         $this->concurrency->add($scheduled);
 
@@ -67,6 +70,7 @@ final class Curl implements Transport
             Period::second(1),
             static fn() => null,
             false,
+            null,
         );
     }
 
@@ -84,6 +88,7 @@ final class Curl implements Transport
             $this->timeout,
             $this->heartbeat,
             $this->disableSSLVerification,
+            $this->proxy,
         );
     }
 
@@ -105,6 +110,7 @@ final class Curl implements Transport
                 default => \Closure::fromCallable($heartbeat),
             },
             $this->disableSSLVerification,
+            $this->proxy,
         );
     }
 
@@ -123,6 +129,23 @@ final class Curl implements Transport
             $this->timeout,
             $this->heartbeat,
             true,
+            $this->proxy,
+        );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    public function proxy(Url $proxy): self
+    {
+        return new self(
+            $this->headerFactory,
+            $this->io,
+            $this->concurrency,
+            $this->timeout,
+            $this->heartbeat,
+            $this->disableSSLVerification,
+            $proxy,
         );
     }
 }